adding a whole bunch of pics

+++ Knowledge_Is_Power/ThunderExplode.java
@@ -55,7 +55,7 @@ public class ThunderExplode extends Bullet
         }else{
             turn(fire_rotation);
         }
-        move(move_speed);
+        move((int)(move_speed));
         setRotation(0);
     }
     public void gifAnimator(){
